+++ src/t8_schemes/t8_default/t8_default_prism/t8_default_prism_cxx.cxx
@@ -484,7 +484,8 @@ t8_default_scheme_prism_c::t8_element_reference_coords (const t8_element_t
                                                         *elem,
                                                         const double
                                                         *ref_coords,
-                                                        const int num_coords,
+                                                        const size_t
+                                                        num_coords,
                                                         double *out_coords)
   const
 {
