+++ src/t8_schemes/t8_default/t8_default_tri/t8_default_tri_cxx.cxx
@@ -580,7 +580,7 @@ t8_default_scheme_tri_c::t8_element_reference_coords (const t8_element_t
                                                       *elem,
                                                       const double
                                                       *ref_coords,
-                                                      const int num_coords,
+                                                      const size_t num_coords,
                                                       double *out_coords)
   const
 {
